Add files via upload
this project is called NeTA, meaning Never Travel Alone. it is a project that was part of my first-semester assessment. it is an android application that connects people that want to have a ride but have insufficient money, therefore, they can use the app and share the ride with someone else that may be willing to travel to a similar location.